Move teams components to teams folders
